MFile: Form1.cs AFunction: Delete_Click() Brief: Added button to delete last entered digit or operator.
MFunction: toPowerOfX_Click()
Brief: Only possible to use this if there is not a operator as last char or empty space.

MFile: Form1.Designer.cs
Brief: Added Delete button. Moved Clear,Modulo,Division buttons to the left. Moved On/Off button to place where it could be in the future.