MFile: Form1.cs MFunction: appendDigit() Brief: Now instead of rewriting the firstnumber_str, now it add up to it when newNumberMode is enabled.
MFunction: operatorButton_Click()
Brief: Now if newNumberMode is true and first char is minus sign it adds to the number.

MFunction: Result_Click()
Brief: If firstnumber_str has minus sign as first char it will remove the sign and set bool variable to true so later it will multiply by -1.